feat: implement object-driven commands and help
